feat: collaborations implemented

+++ src/api/collaborations/handler.js
@@ -48,6 +48,7 @@ class CollaborationsHandler {
 
       const { playlistId, userId } = request.payload
 
+      await this._collaborationsService.verifyCollaborationAccess(playlistId, userId)
       await this._collaborationsService
         .deleteCollaboration(playlistId, userId)
